Python renaming packages from google-cloud to opentelemetry
