--- README.md
@@ -1,19 +1,0 @@
-# opentelemetry-sqlcommenter
-SQLCommenter components for various languages
-
-[Documentation](https://google.github.io/sqlcommenter/)
-
-Contains all the various `sqlcommenter-*` implementations.
-
-- [X] [Python](python/sqlcommenter-python/README.md)
-    - [X] [django](python/sqlcommenter-python/README.md#django)
-    - [X] [psycopg2](python/sqlcommenter-python/README.md#psycopg2)
-    - [X] [sqlalchemy](python/sqlcommenter-python/README.md#sqlalchemy)
-- [X] [Java](java/sqlcommenter-java/README.md)
-    - [X] [Hibernate](java/sqlcommenter-java/README.md#hibernate)
-    - [X] [Spring+Hibernate](java/sqlcommenter-java/README.md#spring-hibernate)
-- [X] Ruby
-    - [X] [Rails](ruby/sqlcommenter-ruby/sqlcommenter_rails/README.md)
-- [X] [Node.js](nodejs/sqlcommenter-nodejs/README.md)
-    - [X] [Knex.js](nodejs/sqlcommenter-nodejs/packages/sqlcommenter-knex/README.md)
-    - [X] [Sequelize.js](nodejs/sqlcommenter-nodejs/packages/sqlcommenter-sequelize/README.md)
