+++ docs/python/django/aws.md
@@ -17,25 +17,25 @@ This guide will help you add [sqlcommenter](https://github.com/open-telemetry/op
 | Steps                     | Resource                                                                   |
 | ------------------------- | -------------------------------------------------------------------------- |
 | Python on AWS             | <https://aws.amazon.com/getting-started/projects/deploy-python-application/> |
-| google-cloud-sqlcommenter | <https://pypi.org/project/google-cloud-sqlcommenter>                         |
+| opentelemetry-sqlcommenter | <https://pypi.org/project/opentelemetry-sqlcommenter>                         |
 | Django 2.X                | <https://docs.djangoproject.com/en/stable/faq/install>                       |
 | Python 3.X                | <https://www.python.org/downloads/>                                          |
 
 ## Addition to your code
 
-Firstly, please install [google-cloud-sqlcommenter](README.md#installation).
+Firstly, please install [opentelemetry-sqlcommenter](README.md#installation).
 
 For any Django deployment, we can just edit your settings.py file and update the `MIDDLEWARE` section
 with
 ```python
 MIDDLEWARE = [
-  'google.cloud.sqlcommenter.django.middleware.SqlCommenter',
+  'opentelemetry.sqlcommenter.django.middleware.SqlCommenter',
   ...
 ]
 ```
 
 >If any middleware execute database queries (that you'd like commented by SqlCommenter), those middleware MUST appear after
-'google.cloud.sqlcommenter.django.middleware.SqlCommenter'
+'opentelemetry.sqlcommenter.django.middleware.SqlCommenter'
 
 ## References
