+++ python/sqlcommenter-python/opentelemetry/sqlcommenter/version.py
@@ -1,12 +1,10 @@
-#!/usr/bin/python
-#
 # Copyright The OpenTelemetry Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#      http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__import__('pkg_resources').declare_namespace(__name__)
+__version__ = "0.1.0"
