+++ python/sqlcommenter-python/setup.cfg
@@ -1,5 +1,5 @@
 [flake8]
-exclude = build,.git,.tox
+exclude = build.git.tox
 max-line-length = 119
 
 [isort]
@@ -9,3 +9,40 @@ include_trailing_comma = true
 known_first_party = sqlcommenter
 line_length = 79
 multi_line_output = 5
+
+[metadata]
+name=opentelemetry_sqlcommenter
+version=attr: opentelemetry.sqlcommenter.version.__version__
+author=OpenTelemetry Authors
+author_email=cncf-opentelemetry-contributors@lists.cncf.io
+description=(Augment SQL statements with meta information about frameworks and the running environment.)
+long_description = file: README.md
+long_description_content_type=text/markdown
+license=BSD
+classifiers=
+    Development Status :: 4 - Beta
+    Environment :: Web Environment
+    Intended Audience :: Developers
+    License :: OSI Approved :: BSD License
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Topic :: Utilities
+    Framework :: Django
+    Framework :: Django :: 2.1
+    Framework :: Django :: 2.2
+
+[options]
+python_requires = >=3
+packages = find:
+
+[options.extras_require]
+django=django>=1.11
+psycopg2=psycopg2
+sqlalchemy=sqlalchemy
+opencensus=opencensus;
+opentelemetry=opentelemetry-api~=1.0
+
