+++ python/sqlcommenter-python/setup.py
@@ -14,48 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
+from setuptools import setup
 
-from setuptools import find_packages, setup
-
-
-def read_file(filename):
-    with open(os.path.join(os.path.dirname(__file__), filename)) as file:
-        return file.read()
-
-
-setup(
-    name='google-cloud-sqlcommenter',
-    version='2.0.0',
-    author='Google Developers',
-    author_email='sqlcommenter@googlegroups.com',
-    description=('Augment SQL statements with meta information about frameworks and the running environment.'),
-    long_description=read_file('README.md'),
-    long_description_content_type='text/markdown',
-    license='BSD',
-    packages=find_packages(exclude=['tests']),
-    extras_require={
-        'django': ['django >= 1.11'],
-        'flask': ['flask'],
-        'psycopg2': ['psycopg2'],
-        'sqlalchemy': ['sqlalchemy'],
-        'opencensus': ['opencensus'],
-        'opentelemetry': ["opentelemetry-api ~= 1.0"],
-    },
-    classifiers=[
-        'Development Status :: 4 - Beta',
-        'Environment :: Web Environment',
-        'Intended Audience :: Developers',
-        'License :: OSI Approved :: BSD License',
-        'Operating System :: OS Independent',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
-        'Programming Language :: Python :: 3.8',
-        'Programming Language :: Python :: 3.9',
-        'Topic :: Utilities',
-        'Framework :: Django',
-        'Framework :: Django :: 2.1',
-        'Framework :: Django :: 2.2',
-    ],
-)
+setup()
