+++ python/sqlcommenter-python/tests/generic/test_opencensus.py
@@ -17,7 +17,7 @@
 from unittest import TestCase
 
 import six
-from google.cloud.sqlcommenter.opencensus import get_opencensus_values
+from opentelemetry.sqlcommenter.opencensus import get_opencensus_values
 
 from ..compat import mock
 from ..opencensus_mock import mock_opencensus_tracer
@@ -25,7 +25,7 @@ from ..opencensus_mock import mock_opencensus_tracer
 
 class OpenCensusTests(TestCase):
     def test_not_installed(self):
-        with mock.patch('google.cloud.sqlcommenter.opencensus.execution_context', new=None):
+        with mock.patch('opentelemetry.sqlcommenter.opencensus.execution_context', new=None):
             with six.assertRaisesRegex(self, ImportError, 'opencensus is not installed'):
                 get_opencensus_values()
