+++ python/sqlcommenter-python/tests/sqlalchemy/tests.py
@@ -17,7 +17,7 @@
 from unittest import TestCase
 
 import sqlalchemy
-from google.cloud.sqlcommenter.sqlalchemy.executor import BeforeExecuteFactory
+from opentelemetry.sqlcommenter.sqlalchemy.executor import BeforeExecuteFactory
 
 from ..compat import mock, skipIfPy2
 from ..opencensus_mock import mock_opencensus_tracer
@@ -82,7 +82,7 @@ class Tests(SQLAlchemyTestCase):
     @skipIfPy2
     def test_both_opentelemetry_and_opencensus_warn(self):
         with mock.patch(
-            "google.cloud.sqlcommenter.sqlalchemy.executor.logger"
+            "opentelemetry.sqlcommenter.sqlalchemy.executor.logger"
         ) as logger_mock, mock_opencensus_tracer(), mock_opentelemetry_context():
             self.assertSQL(
                 "SELECT 1; /*traceparent='00-000000000000000000000000deadbeef-000000000000beef-00',"
@@ -100,27 +100,27 @@ class FlaskTests(SQLAlchemyTestCase):
         'route': '/',
     }
 
-    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('opentelemetry.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_all_data(self, get_info):
         self.assertSQL(
             "SELECT 1; /*controller='c',framework='flask',route='/'*/",
         )
 
-    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('opentelemetry.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_framework_disabled(self, get_info):
         self.assertSQL(
             "SELECT 1; /*controller='c',route='/'*/",
             with_framework=False,
         )
 
-    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('opentelemetry.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_controller_disabled(self, get_info):
         self.assertSQL(
             "SELECT 1; /*framework='flask',route='/'*/",
             with_controller=False,
         )
 
-    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('opentelemetry.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_route_disabled(self, get_info):
         self.assertSQL(
             "SELECT 1; /*controller='c',framework='flask'*/",
